requirements.txt
Flask
python-dotenv
Flask-Cors
google-generativeai
pandas